Updated for Ubuntu 18.04. That tar command is bad
The tar command at the end needs a lot of work. It should only extract Nagios plugins that don't already exist. Instead it overwrites a few files outside of that path, in addition to doing what it should. I am proud of that subshell, though; just came to me. 

I'm concerned that the whole tar command with its subshell and glob will only work in Bash.